refactor(riscv): colocate loadstore cuda wrappers
