+++ extensions/riscv/circuit/src/load_sign_extend/mod.rs
@@ -7,10 +7,5 @@ pub use byte::*;
 pub use halfword::*;
 pub use word::*;
 
-#[cfg(feature = "cuda")]
-mod cuda;
-#[cfg(feature = "cuda")]
-pub use cuda::*;
-
 #[cfg(test)]
 mod tests;
