--- extensions/riscv/circuit/src/loadstore/cuda.rs
@@ -1,192 +1,0 @@
-use std::{mem::size_of, sync::Arc};
-
-use derive_new::new;
-use openvm_circuit::{arch::DenseRecordArena, utils::next_power_of_two_or_zero};
-use openvm_circuit_primitives::{
-    bitwise_op_lookup::BitwiseOperationLookupChipGPU, var_range::VariableRangeCheckerChipGPU, Chip,
-};
-use openvm_cuda_backend::{base::DeviceMatrix, prelude::F, GpuBackend};
-use openvm_cuda_common::copy::MemCopyH2D;
-use openvm_stark_backend::prover::AirProvingContext;
-
-use crate::{
-    adapters::{Rv64LoadStoreAdapterCols, Rv64LoadStoreAdapterRecord, RV64_BYTE_BITS},
-    cuda_abi::{
-        loadstore_byte_cuda, loadstore_doubleword_cuda, loadstore_halfword_cuda,
-        loadstore_word_cuda,
-    },
-    loadstore::{
-        aligned::LoadStoreAlignedCoreCols, LoadStoreByteCoreCols, LoadStoreDoublewordCoreCols,
-        LoadStoreRecord, HALFWORD_SELECTOR_WIDTH, WORD_SELECTOR_WIDTH,
-    },
-};
-
-#[derive(new)]
-pub struct Rv64LoadStoreByteChipGpu {
-    pub range_checker: Arc<VariableRangeCheckerChipGPU>,
-    pub bitwise_lookup: Arc<BitwiseOperationLookupChipGPU<RV64_BYTE_BITS>>,
-    pub pointer_max_bits: usize,
-    pub timestamp_max_bits: usize,
-}
-
-impl Chip<DenseRecordArena, GpuBackend> for Rv64LoadStoreByteChipGpu {
-    fn generate_proving_ctx(&self, arena: DenseRecordArena) -> AirProvingContext<GpuBackend> {
-        const RECORD_SIZE: usize = size_of::<(Rv64LoadStoreAdapterRecord, LoadStoreRecord)>();
-        let records = arena.allocated();
-        if records.is_empty() {
-            return AirProvingContext::simple_no_pis(DeviceMatrix::dummy());
-        }
-        debug_assert_eq!(records.len() % RECORD_SIZE, 0);
-
-        let trace_width =
-            Rv64LoadStoreAdapterCols::<F>::width() + LoadStoreByteCoreCols::<F>::width();
-        let trace_height = next_power_of_two_or_zero(records.len() / RECORD_SIZE);
-        let device_ctx = &self.range_checker.device_ctx;
-
-        let d_records = tracing::info_span!("trace_gen.h2d_records")
-            .in_scope(|| records.to_device_on(device_ctx))
-            .unwrap();
-        let d_trace = DeviceMatrix::<F>::with_capacity_on(trace_height, trace_width, device_ctx);
-
-        unsafe {
-            loadstore_byte_cuda::tracegen(
-                d_trace.buffer(),
-                trace_height,
-                &d_records,
-                self.pointer_max_bits,
-                &self.range_checker.count,
-                &self.bitwise_lookup.count,
-                self.timestamp_max_bits as u32,
-                device_ctx.stream.as_raw(),
-            )
-            .unwrap();
-        }
-        AirProvingContext::simple_no_pis(d_trace)
-    }
-}
-
-#[derive(new)]
-pub struct Rv64LoadStoreHalfwordChipGpu {
-    pub range_checker: Arc<VariableRangeCheckerChipGPU>,
-    pub pointer_max_bits: usize,
-    pub timestamp_max_bits: usize,
-}
-
-impl Chip<DenseRecordArena, GpuBackend> for Rv64LoadStoreHalfwordChipGpu {
-    fn generate_proving_ctx(&self, arena: DenseRecordArena) -> AirProvingContext<GpuBackend> {
-        const RECORD_SIZE: usize = size_of::<(Rv64LoadStoreAdapterRecord, LoadStoreRecord)>();
-        let records = arena.allocated();
-        if records.is_empty() {
-            return AirProvingContext::simple_no_pis(DeviceMatrix::dummy());
-        }
-        debug_assert_eq!(records.len() % RECORD_SIZE, 0);
-
-        let trace_width = Rv64LoadStoreAdapterCols::<F>::width()
-            + LoadStoreAlignedCoreCols::<F, HALFWORD_SELECTOR_WIDTH>::width();
-        let trace_height = next_power_of_two_or_zero(records.len() / RECORD_SIZE);
-        let device_ctx = &self.range_checker.device_ctx;
-
-        let d_records = tracing::info_span!("trace_gen.h2d_records")
-            .in_scope(|| records.to_device_on(device_ctx))
-            .unwrap();
-        let d_trace = DeviceMatrix::<F>::with_capacity_on(trace_height, trace_width, device_ctx);
-
-        unsafe {
-            loadstore_halfword_cuda::tracegen(
-                d_trace.buffer(),
-                trace_height,
-                &d_records,
-                self.pointer_max_bits,
-                &self.range_checker.count,
-                self.timestamp_max_bits as u32,
-                device_ctx.stream.as_raw(),
-            )
-            .unwrap();
-        }
-        AirProvingContext::simple_no_pis(d_trace)
-    }
-}
-
-#[derive(new)]
-pub struct Rv64LoadStoreWordChipGpu {
-    pub range_checker: Arc<VariableRangeCheckerChipGPU>,
-    pub pointer_max_bits: usize,
-    pub timestamp_max_bits: usize,
-}
-
-impl Chip<DenseRecordArena, GpuBackend> for Rv64LoadStoreWordChipGpu {
-    fn generate_proving_ctx(&self, arena: DenseRecordArena) -> AirProvingContext<GpuBackend> {
-        const RECORD_SIZE: usize = size_of::<(Rv64LoadStoreAdapterRecord, LoadStoreRecord)>();
-        let records = arena.allocated();
-        if records.is_empty() {
-            return AirProvingContext::simple_no_pis(DeviceMatrix::dummy());
-        }
-        debug_assert_eq!(records.len() % RECORD_SIZE, 0);
-
-        let trace_width = Rv64LoadStoreAdapterCols::<F>::width()
-            + LoadStoreAlignedCoreCols::<F, WORD_SELECTOR_WIDTH>::width();
-        let trace_height = next_power_of_two_or_zero(records.len() / RECORD_SIZE);
-        let device_ctx = &self.range_checker.device_ctx;
-
-        let d_records = tracing::info_span!("trace_gen.h2d_records")
-            .in_scope(|| records.to_device_on(device_ctx))
-            .unwrap();
-        let d_trace = DeviceMatrix::<F>::with_capacity_on(trace_height, trace_width, device_ctx);
-
-        unsafe {
-            loadstore_word_cuda::tracegen(
-                d_trace.buffer(),
-                trace_height,
-                &d_records,
-                self.pointer_max_bits,
-                &self.range_checker.count,
-                self.timestamp_max_bits as u32,
-                device_ctx.stream.as_raw(),
-            )
-            .unwrap();
-        }
-        AirProvingContext::simple_no_pis(d_trace)
-    }
-}
-
-#[derive(new)]
-pub struct Rv64LoadStoreDoublewordChipGpu {
-    pub range_checker: Arc<VariableRangeCheckerChipGPU>,
-    pub pointer_max_bits: usize,
-    pub timestamp_max_bits: usize,
-}
-
-impl Chip<DenseRecordArena, GpuBackend> for Rv64LoadStoreDoublewordChipGpu {
-    fn generate_proving_ctx(&self, arena: DenseRecordArena) -> AirProvingContext<GpuBackend> {
-        const RECORD_SIZE: usize = size_of::<(Rv64LoadStoreAdapterRecord, LoadStoreRecord)>();
-        let records = arena.allocated();
-        if records.is_empty() {
-            return AirProvingContext::simple_no_pis(DeviceMatrix::dummy());
-        }
-        debug_assert_eq!(records.len() % RECORD_SIZE, 0);
-
-        let trace_width =
-            Rv64LoadStoreAdapterCols::<F>::width() + LoadStoreDoublewordCoreCols::<F>::width();
-        let trace_height = next_power_of_two_or_zero(records.len() / RECORD_SIZE);
-        let device_ctx = &self.range_checker.device_ctx;
-
-        let d_records = tracing::info_span!("trace_gen.h2d_records")
-            .in_scope(|| records.to_device_on(device_ctx))
-            .unwrap();
-        let d_trace = DeviceMatrix::<F>::with_capacity_on(trace_height, trace_width, device_ctx);
-
-        unsafe {
-            loadstore_doubleword_cuda::tracegen(
-                d_trace.buffer(),
-                trace_height,
-                &d_records,
-                self.pointer_max_bits,
-                &self.range_checker.count,
-                self.timestamp_max_bits as u32,
-                device_ctx.stream.as_raw(),
-            )
-            .unwrap();
-        }
-        AirProvingContext::simple_no_pis(d_trace)
-    }
-}
