+++ extensions/riscv/circuit/src/loadstore/mod.rs
@@ -1,9 +1,9 @@
-pub mod aligned;
-pub mod byte;
-pub mod common;
-pub mod doubleword;
-pub mod halfword;
-pub mod word;
+pub(crate) mod aligned;
+mod byte;
+pub(crate) mod common;
+mod doubleword;
+mod halfword;
+mod word;
 
 pub use byte::*;
 pub use common::{LoadStoreExecutor, LoadStoreRecord};
@@ -12,11 +12,6 @@ pub use halfword::*;
 pub use word::*;
 
 mod execution;
-
-#[cfg(feature = "cuda")]
-mod cuda;
-#[cfg(feature = "cuda")]
-pub use cuda::*;
 
 #[cfg(feature = "aot")]
 mod aot;
